Separate enable/disable and changing options
Move the enable/disable function to the options panel so that you can
view/change options without enabling/disabling the debug header.

I really wanted to do this with a "menu button" like firebug and
greasemonkey use, but after playing with it for quite a while I gave
that up. The sdk doesn't support the menu button functionality directly
and hacking up something that works like the sdk buttons and supports
adding a menu was more work than I wanted to dive into for this add-on.

Closes #10